backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-multipart>=0.0.6
SQLAlchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic>=1.13.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
